Separate formatting argument parsing
