+++ src/classes/offset_datetime.rs
@@ -8,7 +8,8 @@ use crate::classes::plain_datetime::DateTimeBoundaryUnit;
 use crate::{
     classes::{date::Date, plain_datetime, time::Time, time_delta::TimeDelta},
     common::{
-        fmt::{self, Suffix},
+        fmt,
+        format_args::{self, Suffix},
         instant::{
             extract_instant, parse_instant_arg, parse_timestamp, parse_timestamp_millis,
             parse_timestamp_nanos,
@@ -541,7 +542,7 @@ fn format_iso(
     args: &[PyObj],
     kwargs: &mut IterKwargs,
 ) -> PyReturn {
-    fmt::format_iso(
+    format_args::format_datetime_iso(
         slf.date,
         slf.time,
         cls.state(),
